# Improve performance of floatformat template filter by using lru_cache