I noticed this block of code was quite expensive and could be cached quite easily.

Before:
```python
In [1]: from django.template.defaultfilters import floatformat
In [2]: f = 1 / 3
In [3]: %timeit floatformat(f)
19.4 µs ± 2.2 µs per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

After:
```python
In [3]: %timeit floatformat(f)
11.7 µs ± 167 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

Not major, but it adds up if this method is called repeatedly.