@sir-sigurd 2017-12-24
That's not fair benchmark, you should try:
```
%timeit floatformat(random.random())
```