@orf 2017-12-24
How come? And wouldn't the overhead of `random.random()` affect the results though? I mean, it's roughly the same speed with `random.random()`:
```
In [7]: %timeit floatformat(random.random())
20.3 µs ± 516 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

But the optimization is to use a cache, in case it's being called repetitively with the same inputs on a page.