@sir-sigurd 2017-12-24
> And wouldn't the overhead of random.random() affect the results though?

```
In [19]: %timeit random.random()
94.8 ns ± 1.1 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [20]: %timeit floatformat(1/3)
29.6 µs ± 996 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [22]: (94.8*10**-9) / (29.6*10**-6)
Out[22]: 0.0032027027027027024
```
So overhead could be neglected.

> in case it's being called repetitively with the same inputs on a page.

How common is this case?