@felixxm 2017-12-25
Agreed. I don't think it's a common use case.