@timgraham 2017-12-26
I'd also agree and say that if this is a bottleneck for a particular project and they know their inputs are such that this will always help and not hinder, they should use a custom template tag.