@orf 2017-12-26
Fair enough, I understand. However there is a small code improvement to
remove the multiple calls to append/the while loop that might be worth
including, I will do that as a separate PR if anyone is interested.